#3987 [Feature Request] Synchronize Entity Type https://github.com/adempiere/adempiere/issues/3987 - Move the abstract domain model to core org.adempiere.core.domains.models
