+++ base/src/org/adempiere/core/domains/models/X_WM_InOutBoundLine.java
@@ -16,7 +16,7 @@
  * or https://github.com/adempiere/adempiere/blob/develop/license.html        *
  *****************************************************************************/
 /** Generated Model - DO NOT CHANGE */
-package org.eevolution.wms.model;
+package org.adempiere.core.domains.models;
 
 import java.math.BigDecimal;
 import java.sql.ResultSet;
