+++ org.eevolution.warehouse/src/main/java/base/org/eevolution/wms/model/MWMSection.java
@@ -28,13 +28,14 @@
  **********************************************************************/
 package org.eevolution.wms.model;
 
+import org.adempiere.core.domains.models.X_WM_Section;
+import org.compiere.model.Query;
+import org.compiere.util.CLogger;
+
 import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
-
-import org.compiere.model.Query;
-import org.compiere.util.CLogger;
 
 /**
  * Class Model Warehouse Section
