+++ org.eevolution.warehouse/src/main/java/base/org/eevolution/wms/model/MWMSectionType.java
@@ -28,12 +28,13 @@
  **********************************************************************/
 package org.eevolution.wms.model;
 
+import org.adempiere.core.domains.models.X_WM_Section_Type;
+import org.compiere.model.Query;
+import org.compiere.util.CLogger;
+
 import java.sql.ResultSet;
 import java.util.List;
 import java.util.Properties;
-
-import org.compiere.model.Query;
-import org.compiere.util.CLogger;
 
 /**
  * Class Model Warehouse Section Type
