+++ org.eevolution.warehouse/src/main/java/base/org/eevolution/wms/model/MWMStrategyDetail.java
@@ -28,10 +28,11 @@
  **********************************************************************/
 package org.eevolution.wms.model;
 
+import org.adempiere.core.domains.models.X_WM_Strategy_Detail;
+import org.compiere.util.CLogger;
+
 import java.sql.ResultSet;
 import java.util.Properties;
-
-import org.compiere.util.CLogger;
 
 /**
  * Class Model Warehouse Strategy Detail
