+++ org.spin.store/src/main/java/org/spin/store/model/MWPaymentMethodAllocation.java
@@ -15,6 +15,8 @@
  ************************************************************************************/
 package org.spin.store.model;
 
+import org.adempiere.core.domains.models.X_C_PaymentMethodAllocation;
+
 import java.sql.ResultSet;
 import java.util.Properties;
